Separate avro dependencies from the others
