+++ stdlib/messaging/kafka/src/main/java/org/ballerinalang/messaging/kafka/nativeimpl/producer/SendIntValues.java
@@ -18,9 +18,7 @@
 
 package org.ballerinalang.messaging.kafka.nativeimpl.producer;
 
-import org.apache.avro.generic.GenericRecord;
 import org.apache.kafka.clients.producer.ProducerRecord;
-import org.ballerinalang.jvm.values.MapValue;
 import org.ballerinalang.jvm.values.ObjectValue;
 import org.ballerinalang.jvm.values.api.BArray;
 import org.slf4j.Logger;
@@ -87,17 +85,6 @@ public class SendIntValues extends Send {
         Long timestampValue = getLongValue(timestamp);
         ProducerRecord<byte[], Long> kafkaRecord = new ProducerRecord<>(topic, partitionValue, timestampValue,
                                                                         key.getBytes(), value);
-        return sendKafkaRecord(kafkaRecord, producer);
-    }
-
-    // ballerina int and AvroRecord
-    public static Object sendIntAvro(ObjectValue producer, long value, String topic, MapValue<String, Object> key,
-                                     Object partition, Object timestamp) {
-        GenericRecord genericRecord = createGenericRecord(key);
-        Integer partitionValue = getIntValue(partition, ALIAS_PARTITION, logger);
-        Long timestampValue = getLongValue(timestamp);
-        ProducerRecord<GenericRecord, Long> kafkaRecord = new ProducerRecord<>(topic, partitionValue, timestampValue,
-                                                                               genericRecord, value);
         return sendKafkaRecord(kafkaRecord, producer);
     }
